Extract scripting reference helpers
